create new class CaloriesFileHandler to handle storage needs for calories list
